Remove unnecessary useCallback statements
Disable react/jsx-handler-names eslintrc flag

Define constants instead of accessing objects attributes constantly

Create hooks.js file
