+++ src/components/Slider/Slider.jsx
@@ -1,108 +1,9 @@
-import React, { useCallback, useRef } from "react";
+import React from "react";
 import PropTypes from "prop-types";
-import {
-  getClientPosition,
-  addEventListeners,
-  DIMENSIONS_MAP,
-  POSITIONS_MAP
-} from "./helpers";
 import { classNames } from "../../utils";
+import SliderHandle from "./SliderHandle.jsx";
+import useSlider from "./hooks";
 import styles from "./styles.css";
-import SliderHandle, { useSliderHandle } from "./SliderHandle.jsx";
-
-function useSlider(double, disabled, axis, min, max, step, handles) {
-  const container = useRef(null);
-
-  if (double) {
-    handles[0].barrier = useCallback((value) => value <= handles[1].value, [
-      handles
-    ]);
-    handles[1].barrier = useCallback((value) => value >= handles[0].value, [
-      handles
-    ]);
-  }
-
-  const handleHooks = handles.map((handle) =>
-    useSliderHandle(
-      container,
-      double,
-      disabled,
-      axis,
-      min,
-      max,
-      step,
-      handle.value,
-      handle.onChange,
-      handle.onDragStart,
-      handle.onDragEnd,
-      handle.barrier
-    )
-  );
-
-  const handleTrackMouseDown = useCallback(
-    (e) => {
-      if (disabled) return;
-
-      e.preventDefault();
-      const clientPos = getClientPosition(e);
-      const rect = container.current.getBoundingClientRect();
-
-      let handleIndex = 0;
-
-      if (double) {
-        const handlePosition = handleHooks.map(
-          (handleHook) =>
-            handleHook.handle.current.getBoundingClientRect()[axis] + 9
-        );
-
-        if (
-          (clientPos[axis] > handlePosition[0] &&
-            clientPos[axis] < handlePosition[1] &&
-            clientPos[axis] - handlePosition[0] >=
-              handlePosition[1] - clientPos[axis]) ||
-          clientPos[axis] >= handlePosition[1]
-        )
-          handleIndex = 1;
-      }
-
-      addEventListeners(
-        handleHooks[handleIndex].handleDrag,
-        handleHooks[handleIndex].handleDragEnd
-      );
-
-      handleHooks[handleIndex].change(
-        clientPos[axis] - rect[POSITIONS_MAP[axis]]
-      );
-
-      if (handleHooks[handleIndex].onDragStart) {
-        handleHooks[handleIndex].onDragStart(e);
-      }
-    },
-    [axis, disabled, double, handleHooks]
-  );
-
-  const valueStyle = {};
-
-  if (double) {
-    valueStyle[POSITIONS_MAP[axis]] =
-      handleHooks[0].position[POSITIONS_MAP[axis]] + "%";
-
-    valueStyle[DIMENSIONS_MAP[axis]] =
-      handleHooks[1].position[POSITIONS_MAP[axis]] -
-      handleHooks[0].position[POSITIONS_MAP[axis]] +
-      "%";
-  } else {
-    valueStyle[DIMENSIONS_MAP[axis]] =
-      handleHooks[0].position[POSITIONS_MAP[axis]] + "%";
-  }
-
-  return {
-    container,
-    handleTrackMouseDown,
-    valueStyle,
-    handleHooks
-  };
-}
 
 const Slider = ({
   double,
