+++ src/components/Slider/styles.css
@@ -2,11 +2,11 @@
   position: relative;
   display: block;
   content: "";
-  width: 18px;
-  height: 18px;
+  width: 1.8rem;
+  height: 1.8rem;
   background-color: var(--main-color);
   border-radius: 50%;
-  box-shadow: 0 1px 1px rgba(0, 0, 0, 0.5);
+  box-shadow: 0 0.1rem 0.1rem rgba(0, 0, 0, 0.5);
   user-select: none;
   cursor: pointer;
   box-sizing: border-box;
@@ -16,25 +16,25 @@
   position: relative;
   display: inline-block;
   background-color: var(--btn-disabled-background-color);
-  border-radius: 5px;
+  border-radius: 0.5rem;
   user-select: none;
   box-sizing: border-box;
 }
 
 .x.track {
-  width: 200px;
-  height: 10px;
+  width: 20rem;
+  height: 1rem;
 }
 
 .y.track {
-  width: 10px;
-  height: 200px;
+  width: 1rem;
+  height: 20rem;
 }
 
 .active {
   position: absolute;
   background-color: var(--btn-background-color);
-  border-radius: 5px;
+  border-radius: 0.5rem;
   user-select: none;
   box-sizing: border-box;
 }
